tlock: Initialize tokens on recieving `FolderChangedMsg`
This fixes a major bug which is when a new user is freshly created, the
app panics because the `Tokens` widget trys to fetch tokens for the
topmost folder, which does not exist. So, rather than initially
fetching, it waits for `FolderChangedMsg` and loads the tokens for the
message (previous implementation for fetching new tokens when folder is
changed). The `FolderChangedMsg` is sent by `DashboardModel` if the
length of the folders is greater than 1.